+++ libbpf-cargo/src/build.rs
@@ -52,7 +52,7 @@ fn extract_version(output: &str) -> Result<&str> {
 /// Extract vendored libbpf header files to a temporary directory.
 ///
 /// Directory and enclosed contents will be removed when return object is dropped.
-#[cfg(not(feature = "novendor"))]
+#[cfg(feature = "default")]
 fn extract_libbpf_headers_to_disk(target_dir: &Path) -> Result<Option<PathBuf>> {
     use std::fs::OpenOptions;
     use std::io::Write;
@@ -69,9 +69,9 @@ fn extract_libbpf_headers_to_disk(target_dir: &Path) -> Result<Option<PathBuf>> 
     Ok(Some(parent_dir))
 }
 
-#[cfg(feature = "novendor")]
-fn extract_libbpf_headers_to_disk(target_dir: &Path) -> Result<Option<PathBuf>> {
-    return Ok(None);
+#[cfg(not(feature = "default"))]
+fn extract_libbpf_headers_to_disk(_target_dir: &Path) -> Result<Option<PathBuf>> {
+    Ok(None)
 }
 
 fn check_clang(debug: bool, clang: &Path, skip_version_checks: bool) -> Result<()> {