+++ libbpf-cargo/CHANGELOG.md
@@ -1,5 +1,8 @@
 Unreleased
 ----------
+- Removed `novendor` feature in favor of having disableable default
+  feature
+- Updated `libbpf-sys` dependency to `1.3.0`
 - Bumped minimum Rust version to `1.66`